requirements.txt
asgiref==3.4.1
autopep8==1.5.7
boto3==1.26.19
botocore==1.29.19
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click @ file:///C:/ci/click_1646038595831/work
colorama @ file:///C:/Windows/TEMP/abs_9439aeb1-0254-449a-96f7-33ab5eb17fc8apleb4yn/croots/recipe/colorama_1657009099097/work
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
decorator==4.4.2
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.7
Django==3.2.6
django-allauth==0.52.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-rest-knox==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dynaconf==3.1.11
gunicorn==20.1.0
hvac==1.0.2
idna==3.4
imageio==2.25.0
imageio-ffmpeg==0.4.8
importlib-metadata==4.0.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib @ file:///C:/b/abs_e60_bwl1v6/croot/joblib_1666298845728/work
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
moviepy==1.0.3
mysqlclient==2.1.1
nltk @ file:///opt/conda/conda-bld/nltk_1645628263994/work
numpy @ file:///C:/b/abs_53f_dbvhzc/croot/numpy_and_numpy_base_1665773185489/work
oauthlib==3.2.2
phonenumbers==8.13.0
Pillow==9.3.0
proglog==0.1.10
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.7.0
pycparser==2.21
pyhcl==0.4.4
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytube==12.1.2
pytz==2021.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
shortuuid==1.0.11
six @ file:///tmp/build/80754af9/six_1644875935023/work
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tqdm @ file:///C:/b/abs_0axbz66qik/croots/recipe/tqdm_1664392691071/work
typing-extensions==3.10.0.0
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.12
validators==0.20.0
whitenoise==5.3.0
win32-setctime==1.1.0
wincertstore==0.2
zipp==3.4.1
